+++ addOns/client/src/test/java/org/zaproxy/addon/client/spider/actions/SubmitFormUnitTest.java
@@ -92,7 +92,7 @@ class SubmitFormUnitTest {
         WebDriver wd = mock(WebDriver.class);
         WebElement form = visibleElement();
         given(wd.findElement(any(By.class))).willReturn(form);
-        given(form.findElements(any(By.class))).willReturn(List.of());
+        given(wd.findElements(any(By.class))).willReturn(List.of());
 
         // When
         boolean result = action.run(waitStrategy, wd);
@@ -104,7 +104,7 @@ class SubmitFormUnitTest {
     }
 
     @Test
-    void shouldFillInputsFromFormBeforeSubmitting() {
+    void shouldFillInputsAndTextAreasFromFormBeforeSubmitting() {
         // Given
         ClientSideComponent component = formComponent(0, "xpath", "//FORM");
         SubmitForm action = new SubmitForm(valueProvider, uri, component);
@@ -113,22 +113,27 @@ class SubmitFormUnitTest {
         given(wd.findElement(any(By.class))).willReturn(form);
         WebElement input1 = visibleInput("inputA", "text");
         WebElement input2 = visibleInput("inputB", "text");
-        given(form.findElements(any(By.class))).willReturn(List.of(input1, input2));
+        WebElement textarea = visibleTextArea("textareaA");
+        given(wd.findElements(any(By.class))).willReturn(List.of(input1, input2, textarea));
         given(valueProvider.getValue(any(), any(), eq("inputA"), any(), any(), any(), any()))
                 .willReturn("value1");
         given(valueProvider.getValue(any(), any(), eq("inputB"), any(), any(), any(), any()))
                 .willReturn("value2");
+        given(valueProvider.getValue(any(), any(), eq("textareaA"), any(), any(), any(), any()))
+                .willReturn("value3");
 
         // When
         boolean result = action.run(waitStrategy, wd);
 
         // Then
         assertThat(result, is(equalTo(true)));
-        InOrder inOrder = inOrder(input1, input2);
+        InOrder inOrder = inOrder(input1, input2, textarea);
         inOrder.verify(input1).clear();
         inOrder.verify(input1).sendKeys("value1");
         inOrder.verify(input2).clear();
         inOrder.verify(input2).sendKeys("value2");
+        inOrder.verify(textarea).clear();
+        inOrder.verify(textarea).sendKeys("value3");
     }
 
     @Test
@@ -139,7 +144,7 @@ class SubmitFormUnitTest {
         WebDriver wd = mock(WebDriver.class);
         WebElement form = visibleElement();
         given(wd.findElement(any(By.class))).willReturn(form);
-        given(form.findElements(any(By.class))).willReturn(List.of());
+        given(wd.findElements(any(By.class))).willReturn(List.of());
         willThrow(RuntimeException.class).given(form).submit();
 
         // When / Then
@@ -194,7 +199,7 @@ class SubmitFormUnitTest {
         WebElement visibleForm = visibleElement();
         ArgumentCaptor<By> byCaptor = ArgumentCaptor.forClass(By.class);
         given(wd.findElement(byCaptor.capture())).willReturn(visibleForm);
-        given(visibleForm.findElements(any(By.class))).willReturn(List.of());
+        given(wd.findElements(any(By.class))).willReturn(List.of());
 
         // When
         boolean result = action.run(waitStrategy, wd);
@@ -247,5 +252,13 @@ class SubmitFormUnitTest {
         given(input.getDomAttribute("name")).willReturn(name);
         given(input.getDomAttribute("type")).willReturn(type);
         return input;
+    }
+
+    private static WebElement visibleTextArea(String name) {
+        WebElement textarea = mock(WebElement.class);
+        given(textarea.isDisplayed()).willReturn(true);
+        given(textarea.getTagName()).willReturn("textarea");
+        given(textarea.getDomAttribute("name")).willReturn(name);
+        return textarea;
     }
 }