+++ addOns/client/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Use new functionality from the browser extension for crawling.
+- Fill text areas while crawling.
 
 ## [0.28.0] - 2026-06-26
 ### Added